BAU: Add ESlint and prettier config to the root of the repo and enable ESlint plugins
Use one config for ESlint and prettier in the root of the repo without the need to duplicate the config files for each lambda. All lambdas are automatically added to the parser by configuring the ESlint project property.

Add a `package.json` file at the root of the repo so ESlint, prettier and the plugins can be installed once and re-used for all lambdas - npm no longer needs to duplicate them in each sub-project and each lambda can now have a slimmer dependency file.

Enable the prettier and TypeScript plugins in ESlint. These plugins were already installed in `package.json` but they weren't used, as they need to be explicitly added to ESlint config.

With the plugins enabled, ESlint highlights prettier rules in the IDE and fixes the errors if the `--fix` flag is used whilst the TypeScript plugin allows ESlint to apply the recommended TypeScript rules.

Update the linting scripts - prettier doesn't need to be run in addition to ESlint as ESlint runs prettier rules with the plugin now in use. In addition, all lambdas can be linted at once with one invocation from the root of the repo, or individually when the linter is invoked from a lambda's root dir.

Remove unused ignore files - files to ignore are already specified in the relevant config files.

Add a prettier ignore file to repos and exclude package-lock.json by default (using the prettier ignore file).