OJ-1984: Enable running all lambda unit tests from the root of the repo
Add jest config to the root of lambda repos so all tests can be run with one command and one coverage report is generated for the whole repo. Tests for an individual lambda can still be run the same as before by cd'ing into the lambda's directory.

Add a base jest config in the root of the repo to avoid duplication of the same settings in each lambda. Add a small jest config to each lambda to set the display name, and override any settings if needed.

Add a `package.json` file at the root of the repo so jest packages can be installed once and re-used for all lambdas - npm no longer needs to duplicate them in each sub-project and each lambda can have a slimmer dependency file.

- Update the testing scripts and add a command to generate a coverage report on demand
- Update jest config to correctly collect coverage
- Use default config values
- Exclude files in the build directory from jest