OJ-1984: Set up npm workspaces for lambda repos
Enable the monorepo config for npm so all packages for all lambdas can be installed at once by running `npm install` at the root of the repo. Other npm commands can be run across the entire repo as well.

npm packages will be installed to the `npm_modules` directory at the root of the repo rather than in individual lambda directories. Common packages used by all lambdas are listed in the root package.json file, and each lambda still has its own package.json file to install additional packages for that lambda.

The SAM template has been adjusted to account for the fact that the root of the repo is the base dir now when running SAM builds, since it needs to refer to the root package.jsonn file. The handler is specified from the root of the repo as well.

Add SAM scripts to the root of the repo.