OJ-1984: Add universal unit test and SonarCloud workflows
Add a unit test workflow that can be run in lambda and frontend repos, and remove the lambda-specific workflow. Remove workflows from the lambda generator since there are no more lambda-specific workflows.

Add a repo scan workflow that can be run in any repo and uses a shared action to perform a SonarCloud scan. SonarCloud scan needs the test coverage results, which are produced by the unit test workflow.

Add an option to the repo generator to trigger workflow initialisation separately.